removed js react demo
