+++ package.json
@@ -10,9 +10,8 @@
     "build:all": "pnpm build:tokens && pnpm build:icons && pnpm build:ui",
     "storybook": "pnpm --filter @bui/ui storybook dev",
     "storybook:build": "pnpm --filter @bui/ui build-storybook",
-    "demo:react": "pnpm --filter @bui/demo-react dev",
+    "demo:react": "pnpm --filter @bui/demo-react-ts dev",
     "demo:svelte": "pnpm --filter @bui/demo-svelte dev",
-    "demo:react-ts": "pnpm --filter @bui/demo-react-ts dev",
     "demo:svelte-ts": "pnpm --filter @bui/demo-svelte-ts dev",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
