--- packages/demo-react-ts/src/App.css
@@ -1,72 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-body {
-  font-family: 'Inter', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
-  background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-  min-height: 100vh;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.app {
-  width: 100%;
-  max-width: 1200px;
-  padding: 2rem;
-  position: relative;
-}
-
-.theme-toggle {
-  position: absolute;
-  top: 1rem;
-  right: 1rem;
-  background: var(--background);
-  /* border: 2px solid rgba(255, 255, 255, 0.3); */
-  border-radius: 50%;
-  width: 48px;
-  height: 48px;
-  cursor: pointer;
-  font-size: 1rem;
-  transition: all 0.3s ease;
-  z-index: 10;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.theme-toggle:hover {
-  /* transform: scale(1.1); */
-}
-
-.theme-toggle:active {
-  /* transform: scale(0.95); */
-}
-
-.main {
-  text-align: center;
-  background: var(--background);
-  backdrop-filter: blur(10px);
-  border-radius: 20px;
-  padding: 4rem 2rem;
-  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.1);
-}
-
-h1 {
-  font-size: 3.5rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin-bottom: 1rem;
-  letter-spacing: -0.02em;
-}
-
-p {
-  font-size: 1.25rem;
-  color: var(--text-secondary);
-  margin-bottom: 3rem;
-  font-weight: 400;
-  line-height: 1.6;
-}
