--- packages/demo-react/vite.config.js
@@ -1,16 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vitejs.dev/config/
-export default defineConfig({
-  plugins: [react()],
-  resolve: {
-    alias: {
-      '@bui/ui': '../../ui',
-      '@bui/tokens': '../../tokens'
-    }
-  },
-  optimizeDeps: {
-    include: ['lit']
-  }
-})
