+++ packages/demo-react/vite.config.ts
